third_party/WebKit/LayoutTests/inspector-protocol/input/emulateTouchFromMouseEvent.html

<html>
<head>
<script type="text/javascript" src="../../http/tests/inspector-protocol/inspector-protocol-test.js"></script>
<script>

window.addEventListener("touchstart", logEvent);
window.addEventListener("touchend", logEvent);
window.addEventListener("touchmove", logEvent);
window.addEventListener("touchcancel", logEvent);

function logEvent(event)
{
    event.preventDefault();
    log("-----Event-----");
    log("type: " + event.type);
    if (event.shiftKey)
        log("shiftKey");
    log("----Touches----");
    for (var i = 0; i < event.touches.length; i++) {
        var touch = event.touches[i];
        log("id: " + i);
        log("pageX: " + touch.pageX);
        log("pageY: " + touch.pageY);
        log("webkitRadiusX: " + touch.webkitRadiusX);
        log("webkitRadiusY: " + touch.webkitRadiusY);
        log("webkitRotationAngle: " + touch.webkitRotationAngle);
        log("webkitForce: " + touch.webkitForce);
    }
    evaluateInFrontend("onEvent();");
}

function test()
{
    var events = [
        {
            "type": "mousePressed",
            "button": "left",
            "clickCount": 1,
            "x": 100,
            "y": 200
        },
        {
            "type": "mouseMoved",
            "button": "left",
            "clickCount": 1,
            "x": 100,
            "y": 250
        },
        {
            "type": "mouseReleased",
            "button": "left",
            "clickCount": 1,
            "x": 100,
            "y": 250
        },
        {
            "type": "mouseMoved",
            "button": "none",
            "modifiers": 8, // shift
            "x": 50,
            "y": 150
        },
        {
            "type": "mousePressed",
            "button": "left",
            "clickCount": 1,
            "x": 100,
            "y": 200
        },
        {
            "type": "mouseMoved",
            "button": "left",
            "clickCount": 1,
            "x": 100,
            "y": 250
        },
        {
            "type": "mouseReleased",
            "button": "left",
            "clickCount": 1,
            "x": 100,
            "y": 250
        }
    ];

    var time = Number(new Date()) / 1000;
    var index = 0;

    function sendNextEvent() {
        if (index === events.length) {
            maybeFinish();
            return;
        }
        var event = events[index];
        event.timestamp = time + index;
        index++;
        InspectorTest.sendCommand("Input.emulateTouchFromMouseEvent", event, checkResponse);
    }

    function checkResponse(msg)
    {
        if (msg.error)
            InspectorTest.log("Error: " + msg.error.message);
        sendNextEvent();
    }

    const expectedEventCount = 6;
    var eventCount = 0;
    window.onEvent = function()
    {
        eventCount++;
        maybeFinish();
    };

    function maybeFinish()
    {
        if (index === events.length && eventCount === expectedEventCount)
            InspectorTest.completeTest();
    }

    InspectorTest.sendCommand("Page.enable", {}, function() {
        InspectorTest.sendCommand("Page.startScreencast", {}, sendNextEvent);
    });
}

</script>
</head>
<body onload="runTest()">
</body>
</html>
