third_party/WebKit/LayoutTests/inspector-protocol/input/emulateTouchFromMouseEvent-expected.txt

-----Event-----
type: touchstart
----Touches----
id: 0
pageX: 100
pageY: 200
webkitRadiusX: 11.5
webkitRadiusY: 11.5
webkitRotationAngle: 0
webkitForce: 1
-----Event-----
type: touchmove
----Touches----
id: 0
pageX: 100
pageY: 250
webkitRadiusX: 11.5
webkitRadiusY: 11.5
webkitRotationAngle: 0
webkitForce: 1
-----Event-----
type: touchend
----Touches----
-----Event-----
type: touchstart
----Touches----
id: 0
pageX: 100
pageY: 200
webkitRadiusX: 11.5
webkitRadiusY: 11.5
webkitRotationAngle: 0
webkitForce: 1
-----Event-----
type: touchmove
----Touches----
id: 0
pageX: 100
pageY: 250
webkitRadiusX: 11.5
webkitRadiusY: 11.5
webkitRotationAngle: 0
webkitForce: 1
-----Event-----
type: touchend
----Touches----
